libc: Add _FORTIFY_SOURCE_STRICT_MEMCPY
Introduce a new compile time option _FORTIFY_SOURCE_STRICT_MEMCPY.
This is intended to detect misuses of memcpy() where copied data
will fall outside of a struct entry.

Consider the following code:

  struct V { int a; char buf1[10]; int b; char buf2[10]; } var;
  int *q = &var.a;
  int *r = &var.b;
  // initialize a and buf1
  memcpy(r, q, sizeof(int) + (10*sizeof(char)));

q is a pointer to b within the struct, and the memcpy is attempting
to write beyond b and into buf2. Currently this is permitted with the
current FORTIFY_SOURCE settings, as there's lots of code out there
which has this behavior.

Enabling _FORTIFY_SOURCE_STRICT_MEMCPY will block this behavior
and stop writes from extending beyond the element of the struct being
modified. In this case, it will detect the overflow from b into buf2.

Enable _FORTIFY_SOURCE_STRICT_MEMCPY just for libc. Enabling it for all
of Android causes the system to refuse to boot.

Disable _FORTITY_SOURCE_STRICT_MEMCPY for libc_gdtoa. The gdtoa library
does some nasty, nasty stuff with structures, and my attempt to fix it
fell flat.

My long term hope is to eventually get all of Android running with
_FORTIFY_SOURCE_STRICT_MEMCPY enabled, but for now, I'm happy with
just libc.

Change-Id: I4b014a341108dc1fe25c1e972c5bf86ab2bb59ef