Update patch set 1
Patch Set 1:

thanks!

note that if it's more convenient for you, there's no reason this file needs to be in bionic. you could argue that there's a reason for the tools to be here --- because they need to match the tzcode. but it might make more sense if the bionic and icu tzdata could be updated atomically?

(but the trade-off is "more work now, less work in future".)

Patch-set: 1