Update patch set 4
Patch Set 4:

This is the simple implementation without the workarounds to cope with Clang. Any bugs it can detect will be detected at compile-time, so compiling with GCC *at all* is enough, whether or not it's the code that's shipped.

Once the Clang bugs are fixed (the issue with _RENAME breaking fortifying and the other with bos(null)), the only necessary workaround would be a single public function for aborting at runtime across these fortify functions due to lack of __attribute__((error(...)).

Patch-set: 4