Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

The DIRTY check for the property serial was to deal with the serial number tracking the multi-byte update to the property content. It was a red herring. I am satisfied that we needed to change the memory ordering from do-not-care to acquired so that all accesses that follow are fulfilled after reading the serial.

Patch-set: 2
Label: Code-Review=+1
Label: Verified=+1