Update patch set 1
Patch Set 1: Verified-1

On a nexus 9, this loses in most cases. Especially for small copies, it loses big (the new version can be as bad as half the speed of the old version).

What device was used to benchmark this?

I don't have a cortex device on hand to measure this with at the moment to see if I see the same issue.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1