Update patch set 1
Patch Set 1:

For what it's worth, so far the gains seem to outweigh the losses, but I think you guys would prefer the more consistent results of the current code. I will test this with double linefill enabled tonight and see if it behaves any different for A9 revisions where that is safely enabled. My test systems are OMAP4460, which is "r3p1-50rel0", so it can't be enabled completely safe on it (memtester in particular will spit out a couple errors after running a while). What baffles me most about this is I have never touched assembly in my entire life until last night! How on earth did I manage to find something like this. I only discovered the ldmia vs ldrd thing by digging into the memchr commit that's up for review currently. I also asked the Linaro guys about if the 'good' of this behavior could be leveraged without the 'bad', and was told that it would have an impractical effect on code size; any further thoughts on that?

Patch-set: 1