Fixed fpmath long double bit fields
fpmath.h in Android's libm had an incorrect structure for long double
type. It stated manh (48 bits) and manl (64 bits) while in fact manh is
just a padding with garbage and manl is an actual mantissa (according to
IEEE 754-2008 and x86_64 ABI).

That is, manh, or padding, should come before the 'sign' bit and not as
a part of mantissa. Because of this misplacement freeBSD's logbl
returned incorrect value, and the NDK device test 'math' failed.

The proposed fix change the position of manh, moving it to the end of
the structure (since it's little-endian architecture) and renaming
'manh' to 'padding'.

Change-Id: I508ce44bad895c28ea46823199cc9cc7a67fd132
Signed-off-by: Egor.Kochetov <egor.kochetov@intel.com>