Support for Thread Sanitizer for Bionic (Try 8)
To prevent spurious (and hazard generating) interceptions by TSAN for
bionic-to-bionic calls, we need some duplicate entry points for some select
APIs defined in bionic.

For example, if bionic defines a function foo(), we change it to
define __bionic_foo() and create a weak alias foo() so that
foo can be called externally (outside of bionic) as foo().
Internally, we can use either foo() or __bionic_foo().

Instead of stitching each and every call site to foo(), we simply
use  #define foo __bionic_foo and use additional macro hackery to
generate the alias foo.

The total load for this is two new header files and each modified file
requires two #include lines, one at the beginning, and one at the end.
Please refer to bionic_internal_symbols.h for details.

It has also been updated to include Dan's TLS change:
https://android-review.googlesource.com/#/c/167420/

pthread_once was also added, because recent TSAN now detects deadlocks
better (made it very easy to track down!)

Unlike the prior versions of this patch, I moved the required changes to
the .S files to the ENTRY() and END() macros. This managed to reduce the
size of this patch by about a 1/3.

Bug: 179564

Change-Id: I11ab697e24def9a234186629f8146416298f5019