[179564] Support for Thread Sanitizer for Bionic
To prevent spurious (and hazard generating) interceptions by TSAN for
bionic-to-bionic calls, we need is some duplicate entry points for some select
APIs defined in bionic.

For example, if bionic defines a function foo(), we change it to
define __bionic_foo() and create a weak alias foo() so that
foo can be called externally (outside of bionic) as foo() and internally,
we can use either foo() or __bionic_foo().

Instead of stitching each and every call site to foo(), we simply
use  #define foo __bionic_foo and use additional macro hackery to
generate the alias foo.

The total load for this is two new header files and each modified file
requires two #include lines, one at the beginning, and one at the end.
Please refer to bionic_internal_symbols.h for details.

Please note. This patch is NOT YET COMPLETE!

Change-Id: I724fb8d77059d3e01e1c4336d1aa2c0c550bbb80