Update patch set 3
Patch Set 3:

Clang's new SafeStack feature has a function attribute for disabling it, but sadly there's still no equivalent for stack canaries. I asked about it some time ago in #llvm and there wasn't much interest in providing it because it's only useful in a tiny subset of libc.

Patch-set: 3