add two forms of ASLR for secondary stacks
A randomly sized mapping is placed above secondary stacks mapped by
libc, up to 10% as large as the stack on 32-bit and 50% on 64-bit where
virtual memory is plentiful (there's no increase to pss). This makes
exploits involving overflows from the stack unreliable. It also slightly
reduces the utility of a leaked stack address as the location of data
above the stack won't be precisely known. There is no change to the
pthread API as it's accounted for separately from the stack itself. It's
essentially a userspace implementation of grsecurity's RAND_THREADSTACK,
although it triggers a segmentation fault instead of immediate death.

The initial stack pointer is also placed at a random offset within the
initial page, using up to 1% of the stack's memory. This also applies to
stacks allocated outside of libc. The granularity of this randomization
is determined by the minimum stack alignment which is currently assumed
to be 16. This provides up to 8-bit entropy with typical 4096 byte pages
and the default stack size is large enough to get the full benefit. It
makes exploits requiring precise offsets from or into the stack
unreliable and also provides cache coloring, improving performance for
threads with similar data layout on the stack.

Sample program:

    #include <stdio.h>

    void *print_base(void *data) {
        char stack_value;
        printf("%p\n", &stack_value);
        return data;
    }

    int main(void) {
        pthread_t thread;
        for (unsigned i = 0; i < 10; i++) {
            if (pthread_create(&thread, NULL, print_base, NULL)) {
                return 1;
            }
            if (pthread_join(thread, NULL)) {
                return 1;
            }
        }
        return 0;
    }

Output:

0xaa890d8f
0xaa88beaf
0xaa88d83f
0xaa88a07f
0xaa8872ef
0xaa88593f
0xaa881b0f
0xaa88fc9f
0xaa88726f
0xaa885b5f

Change-Id: I505eee371c67af86263210535e091ca7665d3891