Update patch set 1
Patch Set 1:

Chris, the original commit message wasn't mine, so my apologies for the lack of clarity.  This commit is just from CAF (and they've used a similar memcpy for years).  They mention refactoring it to thumb2 compared to their original commit, not compared to what was in AOSP. In the latest patchset, I changed to their message before that (use performance version of memcpy). I also added the benchmarks which I posted in the comment above this.

Patch-set: 1