Update patch set 2
Patch Set 2:

Jake, do you have any idea what is the most common size on android? It would be good if we manage to collect this data and add the most common sizes for each routine to the Bionic benchmarks (just to assert that we are not deoptimising for the common case).

Probably uprobes is the way to go about in order to find out this.

Patch-set: 2
Reviewer: Gerrit User 1072549 <1072549@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0