Providing __kernel_vsyscall support for i386 syscall
In order to make the choice dynamic, the libc is improved for the
arch-x86 specific case. Each syscall jumps to a predefined routine
called syscallN, where N is the number of parameters of the system call.

It's up to this routine to choose the best path. The only guarantee is
that without triggering a call to optimize, the legacy interrupt 0x80 is
still used.

This syscall switching is applicable to both static executable, dynamic
libc library, and the linker.

The system call trampolines address the dynamic aspect of the syscall
changing, as opposed to a PLT approach where the first resolution would
stay forever. The system call can be rerouted at any stage of the
execution, be that in the linker or in a dynamic libc.so. As text
segment has to remain shareable across processes for libc.so, the
trampoline@GOT logic takes care of the routing.

Change-Id: Ic26c3f23570e6e51e282e652a72f13d24abefd5d
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Signed-off-by: Mingwei Shi <mingwei.shi@intel.com>