Providing __kernel_vsyscall support for i386 syscall
Due to use the PIC code, extra register used to as the syscall pass
parameter register, like ebx, is occupied. It have to save and restore
the pass parameter register value, in order to complete the switch,
we must redefine the trampolines.

It's up to this routine to choose the best path. The only guarantee is
that without triggering a call to __kernel_vsyscal, the legacy syscall is
still used. This syscall switching is applicable to both static executable,
dynamic libc library, and the linker.

In order to make the choice dynamic, the libc is improved for the
arch-x86 specific case. Each syscall jumps to a predefined routine
called syscallN, where N is the number of parameters of the system call.
In current Android OS, the maximum value of N is 6.

The system call trampolines address the dynamic aspect of the syscall
changing, as opposed to a PLT approach where the first resolution would
stay forever. The system call can be rerouted at any stage of the
execution, be that in the linker or in a dynamic libc.so. As text
segment has to remain shareable across processes for libc.so, the
trampoline@GOT logic takes care of the routing.

Notice: In order to eleminate the code conflict when apply this patch, the
patch don't contain the update auto-generated by the gensyscall.py script.
After apply this patch, remember to execute the gensyscalls.py and get all
updates of syscalls under the folder "bionic/libc/arch-x86/syscalls"

Change-Id: Ic26c3f23570e6e51e282e652a72f13d24abefd5d
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Signed-off-by: Mingwei Shi <mingwei.shi@intel.com>