Update patch set 1
Patch Set 1:

An external user was trying to use an atexit function to print jemalloc stats, but it wasn't working. The issue is that the function malloc_fini_impl closes stdin, stdout, stderr. However, this is only done to avoid printing extra leaks related to the file handles. The only problem is that we close these even if we aren't using the special malloc debug hooks.

That's why I've made this change. I can't think of any reason that we did this, does anyone else remember some reason we might rely on these closes?

Patch-set: 1
Label: Verified=0