Update patch set 5
Patch Set 5:

It would be better if it used XOR to encrypt/decrypt the stack pointer and link register rather than just acting as a canary. That way, an attacker with an arbitrary write primitive (not just a sequential overflow) wouldn't be able to just ignore the cookie.

Patch-set: 5
Label: Verified=0