Update patch set 4
Patch Set 4:

the belief of the upstream libm owners is that the 'big' case basically never happens. do you have realistic counter examples? if you do, we should look at rewriting this to not be arm64 specific. (as soon as our ld128 support gets in upstream and we have a new enough prebuilt clang, we'll switch libm to LOCAL_CLANG := true anyway, so the aarch64-GCC-specific part of this change is useless.)

Patch-set: 4