Update patch set 2
Patch Set 2:

I've been looking at this. I found a place where some of the performance was lost and tweaked the code to get the performance back. Mostly for small memcpys (>=4 and <= 16).

I'm looking at the few places where there is some difference in speed. Unfortunately, krait processors can have strange variations in benchmarking, but usually only when working with really small memcpys. I'm trying on a few different devices, to clarify if there is any place where there is a degradation, or if it's due to something else.

I'm going to work through these, but it takes a bit of time since I have this script which runs a ton of different values to check performance along a number of axes. I also run into thermal throttling issues too, so this can be a pain.

I do appreciate the work, so don't think I've forgotten about this.

Patch-set: 2
Label: Verified=0