Update patch set 1
Patch Set 1:

I verified that there is no dwarf unwind information for the __restore and __restore_rt. I also verified that the arm information for this exists, and the bionic unit tests pass.

I also looked and verified that the other hand-coded assembler routines do not have any entries in the arm unwind tables.

Patch-set: 1