Update patch set 8
Patch Set 8:

Dimitry makes a good point.

Since this feature is intended for production use and we would like application libraries to make use of safestack eventually, and "bundled"(?) apps like gmail to use it as soon as possible, we should not meddle with system properties. Instead we should unconditionally enable safestack on all devices, making it part of android libc abi for N+.

If there are no objections, I'll remove the system property check.

There are worries about VA pressure on low end devices. Since we can not get good ASLR on 32-bits anyway w/o fragmenting VA space even more, there is an option of only enabling safestack on 64-bit platforms. We can teach clang to ignore -fsanitize=safe-stack on 32-bit android targets, so this would be transparent to the users.

Patch-set: 8