[PATCH] Support for Thread Sanitizer for Bionic - First Part.
To prevent spurious (and hazard generating) interceptions by TSAN for
bionic-to-bionic calls, we need some duplicate entry points for some select
APIs defined in bionic.

For example, if bionic defines a function foo(), we change it to
define __bionic_foo() and create a weak alias foo() so that
foo can be called externally (outside of bionic) as foo().
Internally, we can use either foo() or __bionic_foo().

In this patch, we do the low hanging fruit by adding a few extra assembly
helper macros so that all routines defined in .S files (such as system calls)
all wind up as being named __bionic_foo(), with foo() being a weak alias.

They are still invoked as foo() in rest of bionic -- i.e. the call sites have
not yet been touched

They are:

ENTRY_BASIC(s) -- which is exactly that
END_BASIC(s) -- pairs with above

With this patch ENTRY(foo) and END(foo) define __bionic_foo() (former) and the
alias foo() (the latter).

ALIAS_SYMBOL(mmap64, mmap) has ben changed so that the symbols
__bionic_mmap64() is properly materialized in addition to mmap64()

In addition the following
ENTRRY_NO_DWARF(s)
ENTRY_PRIVATE(s)
ENTRY_PRIVATE_NO_DWARF(s)

do NOT generate __bionic_foo(). Therefore, they are now all paired with the
equivalent END_(s) markers (because we don't want to materialize the weak
alias bar() if we are not going to materialize __bionic_bar()

The call __bionic_clone() needs special handling, because it is already in the
appropriate name for an internal call.()

This change should disrupt absolutely nothing.

Change-Id: Ib9e6078e1cf25135389a6d75a66e37924a9857ec