Update patch set 8
Patch Set 8:

> I want to ask two questions:
 > 1. why interceptions of bionic by TSAN is bad or risky?
 > 2. where are the interceptions happen? If they happen in the
 > caller's code, why not simply disable TSAN when building bionic?

Second part. TSAN is not being built with bionic. TSAN is a user-space library that the user links against in order to catch race conditions and hazards for multithreaded user applications.

TSAN merely sits between the system libc (bionic, in case of android) and the user application.

i.e. TSAN is not being "integrated" with android or bionic.

Patch-set: 8