Fix all 32 bit definitions for brillo.
Brillo doesn't have any history, so it doesn't need to be backwards
compatible with anything. This changes cleans up most of the definitions
so we have a clean 32 bit libc.

I've changed the __LIBC64_HIDDEN__ to __LIBC32_LEGACY_PUBLIC__ to better
reflect what's going on.

Add a __BIONIC_LIBC32_LEGACY__ macro to use instead of continuing to
add further qualifiers all over the code.

Bug: 24918750
Change-Id: I0d9e19477797d1de86505f9feca03d2aa62c52e1