Update patch set 1
Patch Set 1:

The linker's own executable code will be no more random, and I'd expect that there are more than enough ROP gadgets for an attacker in there. Each dlopen'ed library is also going to introduce a separate, even lower entropy (usually 0 bits) area of executable code and even libc itself dynamically loads 2 libraries after the initial linking. So I'm not really sure that doing this is useful.

An alternative would be inserting a random gap at the end of the libc init code to essentially provide a separate random base for the initial libraries and anything else using mmap (assuming care is taken not to call malloc before then). That doesn't cover dlopen'ed libraries either, but that's simple enough if reserving some virtual memory can be tolerated.

I have an implementation of an isolated region for dynamic libraries using internal first-best-fit allocation (128M on 32-bit, 1G on 64-bit - but it could be smaller on 32-bit and might as well be significantly larger on 64-bit) here, FWIW:

https://github.com/copperhead/android_bionic/commit/bdba0a3d153fe51545453d542d084f06bbb651d0

It inserts a random gap up to the same size as the reserved space, so it also covers the linker code. This provides a lot of entropy (15-bit with 128M + up to 128M reserved) and it would sidestep the jemalloc mmap aligned for all dlopen'ed libraries, which is how people have been writing reliable mediaserver exploits for Lollipop & later (jemalloc wipes all the ASLR entropy, so dlopen'ed libs don't have ASLR at all, just some non-determinism from things like threading).

You might not want to waste a lot of virtual memory on 32-bit, but the region can be pretty small while still covering all but the most bloated applications. There's little reason not to do it on 64-bit, and on 64-bit it would make sense to have a similar isolated region for user-created PROT_EXEC memory (and then other code can be changed to make sure it asks for PROT_EXEC up-front rather than using mprotect to make it isolated).

Patch-set: 1
Reviewer: Gerrit User 1069238 <1069238@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0