Update patch set 15
Patch Set 15:

It could be a huge PROT_NONE region on 64-bit. The limit on the number of threads ends up being the VMA count limit, so it might as well waste a lot of virtual memory. You could also just reserve between 0 and 1TiB for the gap and then immediately unmap it, but it seems better to make it smaller and keep it around, otherwise there are ways to compromise it.

I don't think relying on a hint that can be (and is, with PaX) ignored by the kernel per the standard is the right way to go, even if you can guarantee that the area is unmapped (which is impossible due to threading).

It should be assumed that the attacker can trigger a very large allocation, exhausting most of the address space (but not hitting OOM if it's not actually written to). That's an easy way around any randomization based on mmap hints.

Patch-set: 15