Update patch set 16
Patch Set 16:

I've applied Daniel's idea, but made the 64-bit gap a lot larger (up to 4x the stack size). I still feel that it may not be enough as there is a significant probability of finding the system stack at some constant offset from the unsafe stack. Maybe we should look at a kernel support for this at some point. This should be good enough for now.

Patch-set: 16