Update patch set 6
Patch Set 6:

(1 comment)

This is getting awfully complicated.

I would reach out to one or more of the Google 3 folks (m3b, alasdair) to see if we can borrow their algorithm.

It seems quite tricky to get this right.  I might have tried to rely on the fact that futexes report spurious wakeups differently.  Our cond implementation does that; it's OK to destroy the cv before everyone has exited the wait, because we don't retest the cv.  But I still think we lose here if we wake up with EINTR, since spurious wakeups are clearly not allowed for barriers. 

Aside from the destruction issue, it seems to me that you should be able to alternate between two counters for odd and even barriers?

We can deal with the destruction issue using some static or thread-local state outside the barrier itself.  If we had a thread-indexed data structure that told us which barrier, if any, that thread was currently blocked on, destroy could wait for this barrier to be unused.

An alternative would be if the last action in wait decrements a use count, which is read by destroy.  Destroy might have to spin-wait, since the other threads can't notify, I think.

Patch-set: 6