Update patch set 6
Patch Set 6:

Yabin and I talked.

This potentially does a futex_wake on deallocated/reused memory.  Yabin is correct that we already do that elsewhere.

See https://lkml.org/lkml/2014/12/1/250 for a relevant discussion.

Thus effectively all of bionic libc has to live with the assumption that futexes allow spurious wakeups.  The kernel guarantees they don't, but bionic may wake up a random memory location.

So long as we're OK with that convention, I'm OK with it, and will look at Yabin's code more carefully.

Patch-set: 6