Update patch set 7
Patch Set 7:

(1 comment)

I added the context to the commit message.

Moving properties out of libc would break the libc ABI, so it's not an option for this change.

This commit functions well for me, but there is that lingering issue with the debug_prop area that I need to resolve.

Patch-set: 7
Label: Verified=0