libc: implement kernel vdso syscalls for i386
This patch uses __kernel_vsyscall instead of "int 0x80"
as the syscall entry point. AT_SYSINFO points to
an adapter to mask the arch specific difference and gives a
performance boost on i386 architecture.

Notice: In order to eleminate the code conflict when apply this patch, the patch
don't contain the update auto-generated by the gensyscall.py script. After apply
this patch, remember to execute the gensyscalls.py and get all updates of syscalls
under the folder "bionic/libc/arch-x86/syscalls"

Change-ID: Ib340c604d02c6c25714a95793737e3cfdc3fc5d7
Signed-off-by: Mingwei Shi <mingwei.shi@intel.com>