libc: implement kernel vdso syscalls for i386
This patch give the way using __kernel_vsyscall instead of softirq "int 0x80"
as the syscall enter when VDSO mapped. AT_SYSINFO points to
an adapter to mask the arch specific difference and give a
performance boost on i386 architecture.

Due to there are transport process when VDSO initialized. Before that, using
"int 0x80", after using __kernel_vsyscall.

Due to the libc global vdso is used to save the corresponding syscall entry point:
  case 1: libc global is not initialized, use int 0x80
  case 2: libc global is initialized, AT_SYSINFO set, use __kernel_vsyscall__
  case 3: libc global is initialized, AT_SYSINFO not set, use default int 0x80

TODO: move __int80_syscall into libc_init_common.cpp so we can default to that
and lose the conditional.

Notice: In order to eleminate the code conflict when apply this patch, the patch
don't contain the update auto-generated by the gensyscall.py script. After apply
this patch, remember to execute the gensyscalls.py and get all updates of syscalls
under the folder "bionic/libc/arch-x86/syscalls"

Change-ID: Ib340c604d02c6c25714a95793737e3cfdc3fc5d7
Signed-off-by: Mingwei Shi <mingwei.shi@intel.com>