Update patch set 8
Patch Set 8:

Update on the N9 issue, there's a problem with the fact that __pthread_normal_mutex_lock() uses ScopedTrace, which reads the debug.atrace.tags.enableflags property.  

Removing the mutex in this commit is not enough to work around the problem, so the caller of the pthread_mutex_lock() must be something else in bionic.  I'm continuing my look into this issue.

Patch-set: 8
Label: Verified=0