Update patch set 2
Patch Set 2:

actually, why do we still have the lazy initialization of stdio? should we just dump that and initialize stdio eagerly in with the other libc initialization? looks like the play store is full of references to __sinit, but we could at least make it a no-op.

Patch-set: 2
Label: Code-Review=0