Update patch set 1
Patch Set 1:

How about the binary compatibility issue of struct pthread_internal_t,
if it has a new member in the middle? Is it possible to have mixed binary modules compiled with old and new definitions?

Is there anyway to change __emutls_get_address or pthread_once to work?

Patch-set: 1