Update patch set 1
Patch Set 1: Code-Review-1

Really we should abort here since this is a bug in the linker and any libraries the linker uses before initializing properties (the property values will be wrong anyway, so no reason to try to read them).

There are a few other functions too that this check should be added for, but this brings us back to parity to before https://android-review.googlesource.com/#/c/181794/ was submitted.

Patch-set: 1
Label: Code-Review=-1