Update patch set 1
Patch Set 1: Verified+1

This should be the only special case that we'll need.  There are a few other control messages such as selinux.restorecon_recursive and selinux.reload_policy, but their values are written to property files after their functionality is done.  

That said, we could change their functionality to be similar to ctl. if we wanted as I doubt it would be important for anyone to read out the last parameter to these control messages, but that can be handled in separate patches.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1