Revert "Revert "Remove __sinit and __sdidinit.""
This reverts commit c8bae05f3ff9f1c736f7be70fa17d02795d748bb.

We were breaking init (ueventd) because we initialize system properties
before we initialize stdio. But if system property initialization fails --
because init hasn't got that far, for example -- the system property
initialization code will try to log an error. If the error logging code
can't talk to logd -- because init hasn't got that far, for example --
it'll fall back to writing to stderr. And thus we end up touching stdio
data structures before they've been initialized.

This second attempt takes things further by removing the stdio initialization
function altogether. The data structures for stdin/stdout/stderr can be
statically initialized as data, and -- since we already had to give the
atexit implementation a backdoor for stdio -- we can just admit that we
need to clean up stdio, and that we always do so last.

Change-Id: I936b2eb5e88e4ebaf5516121872b71fc88e5609c