Update patch set 2
Patch Set 2:

ps2 just fixes the commit message to explain that it's (probably) the unconditional fopen rather than any logging that caused the problem. (i still don't _know_ because the kernel doesn't unwind init when it crashes, so we had a pc, not a stack. but this explanation makes more sense.)

Patch-set: 2