Update patch set 3
Patch Set 3:

In practice, reading once was tried, and we found three linker/loader messages popping up in the logs with the old timestamps. By reading immediately we resolved that issue. Also, from what I understand of the use case, the act of changing the logcat timestamp can be made when one decides to take a trace; so anytime. In addition, some of the daemons that can be traced start up before the filesystem mounts, so the value changes after /data mounts, so we need to re-read it in again at that point. logd deals with this transition by estimating and correcting the timestamps in-place and _requires_ the switch-over to be done immediately/atomically in the callers. Yes, there is a minor race where an old timestamp entry has yet to reach logd, in practice we get one of those one in 20 tries or so.

Patch-set: 3