Update patch set 1
Patch Set 1:

I must disagree, the whole point of a calloc call is that you get zeroed memory. If the allocation just happens to return extra usable memory, that should not then be non-zero bytes. That means if you got the memory and it had extra space, you'd have to manually zero it out, which just seems wrong.

Patch-set: 1
Label: Verified=0